Add files via upload
This repository contains the source code for a Bus Management System, developed as a part of my university Computer Science and Engineering Honours 1st-year project. The system is implemented using HTML, CSS (responsive design in progress), PHP, and MySQL. It facilitates efficient bus scheduling, reservation, and management through a user-friendly web interface. Feel free to explore the code and contribute to its improvement.